Centralize Warp Lie group helpers
